Added the multiple side panel.
The side panel will changed based off of the current item that is
selected.
If the selected item is a class then the class options show up,
if the selected item is an assignment then then the assignment options
show up.  This will allow easy editing without the changing of windows.
 (which is nice)