Lots of changes
Updated protobuf files so the id is located in the request.
In database server:
Moved all of the managers to the same package.
Moved all of the handlers to the same package.
All managers now only use the protobufs! :)
In Webclient:
Added a bankproblem creator,
correct a mistake in the course editor
also added mathjax (but it does not seem to be working)


Former-commit-id: 8c8e62528a8b952e386ab1a80e2c6f608378152f