Updated the server so that they can hopefully take in submissions.
Solution server now has a submission handler (more specifically a
handler that takes in the update list to build an entire list)
The answer checker passes communications to the solution server.
The answer checker has its own update handler for experiments and
request a solution on the first instance of an update happening.  (we
need to make sure that these can trigger the comparing regardless of
which ones happens first.
Finally when a solution is stored a message is passed to the database.
TODO: have the database act on the passed message.
Also
MERRY CHRISTMAS!

Former-commit-id: 2088a1ef431de61748f13f1c348c80a23ebc8dde