Made a self sufficient login template.
This can be popped up any time the user needs to login.
Hopefully the newly created connection will work even after we remove
this object.
This is also one less iframe that is used (as they are very bloated)
This should also be more secure as the login system creates their own
connection.