Fixed issue with the fake page.
Fixed issues for testing actual code instead of using it for unit tests.
That way dom references work just fine.

There are now three different types.
There is the fake page which just inserts the page.
There is the shadow fake page so dom is separate but javascript is not
Then there is the iframe fake page so that it acts like it really would
in prod because it uses an iframe


Former-commit-id: 064d097c5f9ef2f6a5bd5ae0166cc370455a0d7f