started work on expansion of problems.
What needs to be done.
The assignment breakdown needs to get the grades using all of the
problems then call the expand Columns.

The expand Columns need to create new columns in for every student
using the studentMap to find the row.

The index can be given (or a cell given instead)
then it can use populate grade (as that has been refactored to take in
everything needed to populate the grades!)
And that should create the cells correctly
As long as the maps sent in are correct


Former-commit-id: e05e81bf3f9e419701c18450fe9ae4407e809ffd