First half of auth refactor
Making the auth system actually exist.
NOTE: this does not handle the population of the auth database.  Or
even how it is contained.

Also it does not compile.

What does exist is a web client socket that handles the talking to the
auth server.  (minus port and host name)
Also what exist is a service that works on the existing data and
attempts to